#162 - secured wish route - not tested

+++ src/main/java/application/rest/WishController.java
@@ -2,6 +2,9 @@ package application.rest;
 
 import application.persistence.entity.Wish;
 import application.rest.domain.WishDTO;
+import application.service.response.ServiceResponseStatus;
+import application.service.security.CustomUserDetails;
+import application.service.user.UserService;
 import application.service.wish.WishService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -16,8 +19,22 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
     @Autowired
     private WishService wishService;
 
+    @Autowired
+    private UserService userService;
+
     @RequestMapping(value = "/customers/{customerId}/wishlist", method = RequestMethod.GET)
     public ResponseEntity<?> readUsersWishes(@PathVariable UUID customerId, Pageable pageable) {
+        if (CustomUserDetails.getCurrentUser() == null) {
+            return new ErrorResponseEntity(ServiceResponseStatus.UNAUTHORIZED);
+        }
+//        // has logged user demanded roles
+//        ServiceResponse<Boolean> hasRolesResponse = userService.hasCurrentUserDemandedRoles(
+//                UserRoleEnum.ROLE_ADMIN
+//        );
+        // error
+        if (!CustomUserDetails.getCurrentUser().getId().equals(customerId)) {
+            return new ErrorResponseEntity(ServiceResponseStatus.READ_FORBIDDEN);
+        }
         return getPageResponseEntity(
                 wishService.readCustomersWishes(customerId, pageable)
         );
@@ -31,12 +48,24 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
     }
 
     @RequestMapping(value = "/customers/{customerId}/wishlist/{wishId}", method = RequestMethod.GET)
-    public ResponseEntity<?> readAddress(@PathVariable UUID customerId, @PathVariable Long wishId) {
+    public ResponseEntity<?> readWish(@PathVariable UUID customerId, @PathVariable Long wishId) {
+        if (CustomUserDetails.getCurrentUser() != null) {
+            return new ErrorResponseEntity(ServiceResponseStatus.UNAUTHORIZED);
+        }
+        if (!CustomUserDetails.getCurrentUser().getId().equals(customerId)) {
+            return new ErrorResponseEntity(ServiceResponseStatus.READ_FORBIDDEN);
+        }
         return read(wishId);
     }
 
     @RequestMapping(value = "/customers/{customerId}/wishlist", method = RequestMethod.POST)
     public ResponseEntity<?> createWish(@PathVariable UUID customerId, @RequestBody WishDTO wishDTO) {
+        if (CustomUserDetails.getCurrentUser() != null) {
+            return new ErrorResponseEntity(ServiceResponseStatus.UNAUTHORIZED);
+        }
+        if (!CustomUserDetails.getCurrentUser().getId().equals(customerId)) {
+            return new ErrorResponseEntity(ServiceResponseStatus.READ_FORBIDDEN);
+        }
         wishDTO.setCustomerUid(customerId);
         return create(wishDTO);
     }
@@ -50,6 +79,12 @@ public class WishController extends AbstractDatabaseController<Wish, Long, WishD
 
     @RequestMapping(value = "/customers/{customerId}/wishlist/{productId}", method = RequestMethod.DELETE)
     public ResponseEntity<?> deleteWish(@PathVariable UUID customerId, @PathVariable UUID productId) {
+        if (CustomUserDetails.getCurrentUser() != null) {
+            return new ErrorResponseEntity(ServiceResponseStatus.UNAUTHORIZED);
+        }
+        if (!CustomUserDetails.getCurrentUser().getId().equals(customerId)) {
+            return new ErrorResponseEntity(ServiceResponseStatus.READ_FORBIDDEN);
+        }
         return getSimpleResponseEntity(
                 wishService.removeProductFromWishes(customerId, productId)
         );
